Source prop can include http request headers on android. (#50)
* Source prop can include http request headers on android.

This is in line with the corresponding change for react-native's
Image component in 0.43.
https://github.com/facebook/react-native/pull/12448

* deconflict ImageEvent event names with core RN

make the event names in `ImageEvent#eventNameForType` different from the
ones used by the core React Native `Image` component. This fixes an
issue where `Image` components do not trigger `on*` callbacks because
the event name has been overwritten by `PhotoView`.