Android: Add support for dynamic buildTools/compileSdk/etc versions
The buildTools version can be obtained from `rootProject.ext` on react-native 0.56. This
commit adds support for that and if the variables are not defined falls back to the
versions expected by react-native (on 0.56).

This also adds a way of customizing the fresco library version
used via `rootProject.ext.frescoVersion`.

inspired by: https://github.com/geektimecoil/react-native-onesignal/issues/442#issuecomment-377177886